% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap.R
\name{.niche_overlap}
\alias{.niche_overlap}
\title{.niche_overlap: Estimate ellipsoid overlap}
\usage{
.niche_overlap(ellipsoid_sp1, ellipsoid_sp2, in_ellip_sp1, in_ellip_sp2)
}
\arguments{
\item{ellipsoid_sp1}{An object of class ellipsoid_metadata.}

\item{ellipsoid_sp2}{An object of class ellipsoid_metadata.}

\item{in_ellip_sp1}{A Logical vector}

\item{in_ellip_sp2}{A Logical vector}
}
\description{
Estimate ellipsoid overlap and other metrics, such as the volume
of the intersection.
}
