% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipsoid_data.R
\name{ellipsoid_data}
\alias{ellipsoid_data}
\title{Function to compute ellipsoid metadata.}
\usage{
ellipsoid_data(centroid, sigma, cf = 0.95)
}
\arguments{
\item{centroid}{A vector with coordinates of niche-centroid.}

\item{sigma}{A square matrix}

\item{cf}{Confidence level}
}
\description{
Function to compute ellipsoid metadata. It computes volume,
semi-axes length and coordinates.
}
\examples{
set.seed(111)
mat <- matrix(rnorm(33),ncol=3)
centroid <- colMeans(mat)
sig <- cov(mat)
emdat <-ellipsoid_data(centroid=centroid,sigma =sig, cf=0.95)
}
