% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{ellipsoid_metadata-class}
\alias{ellipsoid_metadata-class}
\alias{ellipsoid_metadata}
\alias{emeta}
\title{S4 classes to organize ellipsoid data objects}
\description{
S4 classes to organize ellipsoid data objects
}
\section{Slots}{

\describe{
\item{\code{centroid}}{Centroid of the ellipsoid}

\item{\code{sigma}}{A square matrix}

\item{\code{axes_length}}{Length of ellipsoid axes}

\item{\code{axes_coordinates}}{Axes coordinates}

\item{\code{volume}}{Elliposoid volume}

\item{\code{cf}}{Confidence level}

\item{\code{radius}}{Radius of the ellipsoid.}
}}

