% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{ellipsoid_stack-class}
\alias{ellipsoid_stack-class}
\alias{ellipsoid_stack}
\alias{stack}
\title{S4 classes to organize ellipsoid data objects}
\description{
S4 classes to organize ellipsoid data objects
}
\section{Slots}{

\describe{
\item{\code{ellipsoids}}{A list with ellipsoid metadata}
}}

