% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypercube.R
\name{hypercube}
\alias{hypercube}
\title{Estimate a hypercube that contains all ellipsoids}
\usage{
hypercube(ellipsoids, n)
}
\arguments{
\item{ellipsoids}{An object of class ellipsoid_metada or ellipsoid_stack.}

\item{n}{Number of points to create a random sample inside the hypercube.}
}
\description{
Build a hypervolume that contains the union of all ellipsoids
in a raster_stack.
}
\details{
Given the union of all ellipspoids, it generates a dataset X
constituted by the environmental information of a random sample of
\emph{n} points inside the hypervolume.
}
\examples{
\dontrun{
set.seed(111)
mat <- matrix(rnorm(33),ncol=3)
centroid <- colMeans(mat)
sig <- cov(mat)
emdat <- overllip::ellipsoid_data(centroid=centroid,sigma =sig, cf=0.95)
rgl::plot3d(hypercube(emdat,10000),col="red")
emdat2 <- overllip::ellipsoid_data(centroid=centroid+0.5,sigma =sig,cf=0.99)
aa <- stack(emdat,emdat2)
rgl::plot3d(hypercube(aa,10000),col="gray70")
¨}
}
