% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_methods.R
\name{in-ellipsoid}
\alias{in-ellipsoid}
\alias{[,ellipsoid_metadata,missing,missing,missing-method}
\alias{[,ellipsoid_metadata,ANY,ANY,ANY-method}
\alias{[,ellipsoid_metadata,RasterStack,ANY,ANY-method}
\title{Check if an environmental combination is inside an ellipsoid}
\usage{
\S4method{[}{ellipsoid_metadata,missing,missing,missing}(x, i, j, ..., drop = FALSE)

\S4method{[}{ellipsoid_metadata,ANY,ANY,ANY}(x, i, j, ..., suitability = TRUE, drop = TRUE)

\S4method{[}{ellipsoid_metadata,RasterStack,ANY,ANY}(x, i, j, ..., suitability = TRUE, drop = TRUE)
}
\arguments{
\item{x}{A numerical matrix with environmental combinations to test.}

\item{i}{Not used}

\item{j}{Not used.}

\item{...}{Additional parameters for class ellipsoid_metadata}

\item{drop}{Not used.}

\item{suitability}{Logical. Show suitability values.}
}
\value{
A named-logical vector. Names are suitability values
}
\description{
Askif an environmental combination is inside
an ellipsoid. This method applies to ellipsoid_metadata object
\code{`[`} operator.
}
\examples{
\dontrun{
#----------------------------------------------------------------------------
# Ellipsoids in 2-D
#----------------------------------------------------------------------------
set.seed(111)
mat <- matrix(rnorm(20),ncol=2)
centroid <- colMeans(mat)
sig <- cov(mat)
emdat <- overllip::ellipsoid_data(centroid=centroid,sigma =sig, cf=0.95)
hc <- overllip::hypercube(emdat,5000)
in_ellip_a <- emdat[hc]
plot(emdat,semiaxes=TRUE,asp=1,xlab="dim1",ylab="dim2")
points(hc[in_ellip_a,],pch=".")
}
}
