% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_overlap.R
\name{niche_overlap}
\alias{niche_overlap}
\title{niche_metrics: Estimate ellipsoid overlap}
\usage{
niche_overlap(ellipsoid_sp1, ellipsoid_sp2, env_points)
}
\arguments{
\item{ellipsoid_sp1}{An object of class ellipsoid_metadata.}

\item{ellipsoid_sp2}{An object of class ellipsoid_metadata.}

\item{env_points}{A numeric matrix with environmental data}
}
\description{
Estimate ellipsoid overlap and other metrics, such as the volume
of the intersection.
}
\examples{
\dontrun{
library(overllip)
set.seed(111)
mat <- matrix(rnorm(33),ncol=3)
centroid <- colMeans(mat)
sig <- cov(mat)
ellipsoid_sp1 <- overllip::ellipsoid_data(centroid=centroid,
                                          sigma = sig, cf=0.95)
ellipsoid_sp2 <- overllip::ellipsoid_data(centroid=centroid-1,
                                          sigma = sig, cf=0.95)
env_points <- overllip::hypercube(stack(ellipsoid_sp1,
                                        ellipsoid_sp2),100000)

n_overlap <- overllip::niche_overlap(ellipsoid_sp1 = ellipsoid_sp1,
                                     ellipsoid_sp2 = ellipsoid_sp2,
                                     env_points = env_points)

}
}
