% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{overlap_metrics-class}
\alias{overlap_metrics-class}
\alias{overlap_metrics}
\title{S4 classes to organize ellipsoid overlap data}
\description{
S4 classes to organize ellipsoid overlap data
}
\section{Slots}{

\describe{
\item{\code{comparison}}{A stack with metadata of the ellipsoids that are compared}

\item{\code{n_hypercube}}{Number of points in the hypercube that contains the
ellipsoids being compared.}

\item{\code{n_ellipsoid_1}}{Number of points in ellipsoid 1}

\item{\code{n_ellipsoid_2}}{Number of points in ellipsoid 2}

\item{\code{n_union}}{Number of points in the union of ellipsoids}

\item{\code{n_intersection}}{Number of points in the intersection of both ellipsoids}

\item{\code{Jackard_index}}{Jackard index}

\item{\code{intersection_volume}}{Approximated volume of the intersection. The approximation is
estimated from Mote Carlo simulation.}

\item{\code{intersection_ids}}{Indexes of points in the intersection.}
}}

