% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_methods.R
\name{plot-ellipsoid}
\alias{plot-ellipsoid}
\alias{plot,ellipsoid_metadata,missing-method}
\alias{plot,stack_metrics,missing-method}
\title{Plot method (S4) for ellipsoid_metadata-class object}
\usage{
\S4method{plot}{ellipsoid_metadata,missing}(
  x,
  y = NA,
  col = NULL,
  lwd_axes = 2,
  lty_axes = 2,
  semiaxes = FALSE,
  add = FALSE,
  ...
)

\S4method{plot}{stack_metrics,missing}(
  x,
  y = NA,
  col = NULL,
  which_overlaps = "all",
  new_window = TRUE,
  lwd_axes = 2,
  lty_axes = 2,
  semiaxes = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class ellipsoid_metadata.}

\item{y}{Not used}

\item{col}{Plot color}

\item{lwd_axes}{Line width for ellipsoid semi-axes}

\item{lty_axes}{Line type for ellipsoid semi-axes}

\item{semiaxes}{Logical. Show semi-axes}

\item{add}{Add plot.}

\item{...}{Arguments to pass to base::plot, rgl::plot3d,rgl::wire3d or to}

\item{which_overlaps}{Plot all overlaps or just the global intersection}

\item{new_window}{Open a new plot device}
}
\value{
A 2-Dimensional or 3-Dimensional plot
rgl::segments3d
}
\description{
plot is a generic function for plotting of R objects.
The function invokes particular [methods] which depend on the [class] of the first argument.
}
\examples{
\dontrun{
set.seed(111)
mat <- matrix(rnorm(20),ncol=2)
centroid <- colMeans(mat)
sig <- cov(mat)
emdat <- overllip::ellipsoid_data(centroid=centroid,sigma =sig, cf=0.95)
plot(emdat,semiaxes=TRUE,asp=1,xlab="dim1",ylab="dim2")
}
}
