% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_methods.R
\name{stack,missing-method}
\alias{stack,missing-method}
\alias{stack,ellipsoid_metadata-method}
\title{Stack method for class missing of the \pkg{overllip} package.}
\usage{
\S4method{stack}{missing}(x, ...)

\S4method{stack}{ellipsoid_metadata}(x, ..., ellipsoid_names = NULL)
}
\arguments{
\item{x}{An object of class ellipsoid_metadata}

\item{...}{Additional objects of class ellipsoid_metadata}

\item{ellipsoid_names}{A character vector with names for each ellipsoid}
}
\description{
Stack method for class missing of the \pkg{overllip} package.

Stack method for class ellipsoid_stack of the \pkg{overllip} package.
}
