% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\docType{class}
\name{stack_metrics-class}
\alias{stack_metrics-class}
\alias{stack_metrics}
\title{S4 classes to organize ellipsoid overlap data}
\description{
S4 classes to organize ellipsoid overlap data
}
\section{Slots}{

\describe{
\item{\code{ellipsoids_metadata}}{A stack with metadata of the ellipsoids that are compared}

\item{\code{env_data}}{A matrix with environmental data values}

\item{\code{n_hypervolume}}{Number of points in the hypervolume that contains the
ellipsoids being compared.}

\item{\code{n_unions}}{Number of points in the union of ellipsoids}

\item{\code{n_ellipsoids}}{Number of points in each ellipsoid}

\item{\code{n_intersections}}{Number of points in the intersection of both ellipsoids}

\item{\code{Jackard_indices}}{Jackard indeces}

\item{\code{intersection_volumes}}{Approximated volume of the intersection. The approximation is
estimated from Mote Carlo simulation.}

\item{\code{geographic_intersection}}{A raster stack of intersections.}

\item{\code{intersection_ids}}{Indexes of points in the intersection.}

\item{\code{global_intersection_volume}}{Volume of the global intersection.}

\item{\code{global_intersection_ids}}{Record ids of the global intersection.}
}}

