% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_overlap.R
\name{stack_overlap}
\alias{stack_overlap}
\title{stack_overlpa: Estimate ellipsoid overlap for an object of class ellipsoid_stack}
\usage{
stack_overlap(
  ellipsoid_stack,
  env_data,
  suitability_differences = FALSE,
  parallel = FALSE,
  ncores
)
}
\arguments{
\item{ellipsoid_stack}{An object of class ellipsoid_stack.}

\item{env_data}{A numeric matrix with environmental data}

\item{suitability_differences}{Logical. Show differences in suitabilities
(only if env_data is of class RasterStack). See details}

\item{parallel}{Run process in parallel}

\item{ncores}{Number of cores for parallel process}
}
\description{
Estimate ellipsoid overlap and other metrics, such as the volume
of the intersection for an ellipsoid_stack.
}
\details{
If env_data is of class RasterStack and if suitability_differences = TRUE.
The geographic projection of intersections will show the differences between
the suitabilities of the ellipsoids being compared. Positive values mean closer
places to the centroid of ellipsoid A and negative values are values closer to the
centroid of ellipsoid B.
}
\examples{
\dontrun{
bionames <- c("bio5","bio6","bio12")
de_centroid <- c(314.4205,141.8564,1767.3282)
names(de_centroid) <- bionames
de_sigma <- matrix(c(913.3171,621.1225,-1834.9222,
                     621.1225,1693.1855,6825.4082,
                     -1834.9222,6825.4082,802340.5206),
                   ncol=3,dimnames = list(bionames,bionames))
dr_centroid <- c(309.829,156.0725,1759.1617)
names(dr_centroid) <- bionames
dr_sigma <- matrix(c(1379.9633,795.3904,-3377.1604,
                     795.3904,2481.8104,20088.3235,
                     -3377.1604,20088.3235,812031.9645),
                   ncol=3,dimnames = list(bionames,bionames))
dy_centroid <- c(328.197,191.2879,1791.8939)
dy_sigma <- matrix(c(286.6837,52.6501,695.4674,
                     52.6501,775.8082,5174.0618,
                     695.4674,5174.0618,351406.8655),
                   ncol=3,dimnames = list(bionames,bionames))
sp_names <- c("D. ecaudata","D. rotundus","D. youngi")
de_edata <- overllip::ellipsoid_data(centroid = de_centroid,
                                     sigma = de_sigma,cf = 0.95)
dr_edata <- overllip::ellipsoid_data(centroid = dr_centroid,
                                     sigma = dr_sigma,cf = 0.95)
dy_edata <- overllip::ellipsoid_data(centroid = dy_centroid,
                                     sigma = dy_sigma,cf = 0.95)
ellipsoid_stack <- overllip::stack(de_edata,dr_edata,dy_edata,
                                   ellipsoid_names = sp_names)

env_data <- overllip::hypercube(ellipsoids = ellipsoid_stack,
                                n = 10000000)
rmat <- overllip::stack_overlap(ellipsoid_stack = ellipsoid_stack,
                                env_data = env_data,
                                parallel = T,ncores = 4)
overllip::plot(rmat)
#----------------------------------------------------------------------------
# Example with raster data
#----------------------------------------------------------------------------
env_rdata <- raster::getData("worldclim",var='bio', res=10,path="~/Desktop")
env_rdata <- env_rdata[[c("bio5","bio6","bio12")]]
rras <- overllip::stack_overlap(ellipsoid_stack = ellipsoid_stack,
                                env_data = env_rdata,
                                suitability_differences =TRUE,
                                parallel = F)
overllip::plot(rras,semiaxes=TRUE)
raster::plot(rras@geographic_intersection)
}
}
